Restructuring m3o/cmd/client-gen phase 1
